+++ ultransc/__init__.py
@@ -1,3 +1,3 @@
 """Python implementation of the ULTRANSC transcription pipeline."""
 
-__version__ = "0.7.0"
+__version__ = "0.8.0"